src/requirements.txt
Flask>=2.3.0
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
geopy==2.2.0
boto3==1.26.0
